+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change log
 
-
 ## NOT YET RELEASED
 
 - add "data duplication" report - prints report that shows which directory trees appear
   both in a research and vault collection.
+
+## 2025-05-07 v1.11.0
+
+- add "old vs new data" report - prints report on total replica size of data objects,
+  grouped by old (not recently modified) vs new (recently modified) data objects
 
 ## 2025-03-10 v1.10.0